moved auto disconnect to the correct folder
